README: New "Test mode" text
It seems Stripe has made some changes to their dashboard. For this readme the main change is the removal of orange "Test mode" bar at the top. I've updated the text in the readme to not mention it anymore.

There's also minor copy and layout changes in the circled section with the keys.

Honestly, it's still super easy to follow the readme, so if you just want to close this PR that's fine 🙂  And also, of course, do verify that the dashboard has changed for you as well. Can also be something they're just testing for a small subset of users, who knows 🙂